Update README.md
This set of pages was developed over time and (admittedly) with help from BING chat and Chat GPT.   After exhaustive searching and difficulties the final resort was to "cheat".   The end product was simply some data management (wrangling) software to produce needed results and calculations quickly.

With each final product, something new is thought of, and so the end result is a dynamic one.    

Cheers!  John