Manual data loader with overlay funcion
This manual loader requires manual data entry.   When full, this notebook can be saved, and a new one created from the template for data storage.   This one is written for cathodes, and has variables to change your cathode (or anode) material.   You can overlay by indivicual cycle, and you can overlay cathode and anode if you want to make the comparison between oppositing electrodes for example.    The dual X-axis was setup for analysis of commercial batteries which are manufactured in large sheets with an average mass loading, but at the same time for research purposes (and when cells are damaged) mass loading may also be important.

Some limitations are you can only do one cycle at a time (you can change this).   Also, if you for some reason don't have mass data the mass capacity will be non-functional and irrelevant.   For the data plots, the usual matplot lib feature exist with changing colors, and markers.

Some additions that I couldnt figure out initially is taking a data plot, then extracting the final data points.  This is setup to extract the final data points (3) on the charge cycle, and the final data points on the disharge cycle (you need to annotate the voltage manually for this, it is set to 2.52, the final voltage is 2.50, which was good for my samples)  This will provide  a final discharge and final charge capacity which can be used elsewhere.