Template data mapping on large format electrode
Based on commercial battery electrode 30 X 10cm, we remove small 15mm punches for sampling.  This is a tool for illustrating results using color, as well as 2D mapping of samples.   This has already worked out really well for discussions with others.    There is manual data entry, but the time pay-off later is well worth it.