Sample mapping of large format electrodes
This is a tool that plots the SOH, SOC, and capacity of 15mm disk electrodes taken from a large format 30X 10cm electrode sheet.  This has already proven extremely useful in meetings and for presentation purpsoses.  The use of color to indicate health, charge, and capacity is also useful.  By measuring the punch out locations (ImageJ was used here) and keeping track of coordinates of all samples, other opposing samples (like in the cathode opposing this anode sheet) can be more accurately compared.  

For future sample harvesting, the graphic also presents to the operator what parts of the sample have already been tested, and some of the results.   At the same time, the Rectangle" function allows the user to map out missing sections, damaged sections or areas of interest.

Black circles are samples that did not work, samples that were not used in electrochemical tests, or simply samples that were lost and serve as a place holder to allow for good record keeping and data managment.  

One journal page per sheet.