Large format electrode sheet sample map
This is a template for creating sample a sample map from large format (30X10cm) electrode sheets. using pre-defined areas and manually entered values for capacity, thickness, etc., a very simple graphic is good for presentation for data of an electrode sheet.    "A picture is worth a thousand words" is an old adage, this one is worth 10,000 in my opinion.

It sill needs some work, it's a bit tedious plotting sample coordinates, and is a bit time consuming but the time saving downstream is much more significant.